Scale service vertically for fewer replicas (#28)
* Remove deprecated stack field

* Increase memory to 768mb

Instances are operating in stores clusters at
~450mb but sometimes passing the requested 512mb.

Also noticed a relevant improvement in performance
(lower number of replicas) when I increased the
memory a bit, so I believe the GC may be causing some
pressure when the memory gets too close to the limit.

* Explicitly allocate half a CPU core

The default CPU being allocated was really
really small, and we were using only 0.12
of a core. This caused the catalog-api-proxy
to be scaled in dozens of replicas, almost
100, so we can avoid having so many instances
by scaling them vertically instead.